Revert "SoC und Energielimit für Zeitpladen"
